+++ core/src/main/java/dev/redstudio/redcore/math/vectors/Vector2F.java
@@ -4,43 +4,39 @@ import dev.redstudio.redcore.annotations.Experimental;
 
 @Experimental
 @SuppressWarnings("unused")
-public class Vector3D implements VectorVector3Operations<Vector3D> {
+public class Vector2F implements VectorVector2Operations<Vector2F> {
 
     public double x;
     public double y;
-    public double z;
 
-    public Vector3D() {
+    public Vector2F() {
         // No need to do anything
     }
 
-    public Vector3D(final double x, final double y, final double z) {
+    public Vector2F(final double x, final double y) {
         this.x = x;
         this.y = y;
-        this.z = z;
     }
 
     // region Assignment
     /*================================ Assignment ================================*/
 
     @Override
-    public Vector3D copy(final Vector3F vector) {
+    public Vector2F copy(final Vector2F vector) {
         x = vector.x;
         y = vector.y;
-        z = vector.z;
 
         return this;
     }
 
     @Override
-    public Vector3D copy(final Vector3D vector) {
+    public Vector2F copy(final Vector2D vector) {
         x = vector.x;
         y = vector.y;
-        z = vector.z;
 
         return this;
     }
-
+    
     // endregion
 
     // region Math
@@ -50,42 +46,38 @@ public class Vector3D implements VectorVector3Operations<Vector3D> {
     /*================================ Addition ================================*/
 
     @Override
-    public Vector3D add(final Vector3F vector) {
+    public Vector2F add(final Vector2F vector) {
         x += vector.x;
         y += vector.y;
-        z += vector.z;
 
         return this;
     }
 
     @Override
-    public Vector3D add(final Vector3D vector) {
+    public Vector2F add(final Vector2D vector) {
         x += vector.x;
         y += vector.y;
-        z += vector.z;
 
         return this;
     }
 
-    //endregion
+    // endregion
 
     // region Subtraction
     /*================================ Subtraction ================================*/
 
     @Override
-    public Vector3D subtract(final Vector3F vector) {
+    public Vector2F subtract(final Vector2F vector) {
         x -= vector.x;
         y -= vector.y;
-        z -= vector.z;
 
         return this;
     }
 
     @Override
-    public Vector3D subtract(final Vector3D vector) {
+    public Vector2F subtract(final Vector2D vector) {
         x -= vector.x;
         y -= vector.y;
-        z -= vector.z;
 
         return this;
     }
@@ -96,55 +88,49 @@ public class Vector3D implements VectorVector3Operations<Vector3D> {
     /*================================ Multiplication ================================*/
 
     @Override
-    public Vector3D multiply(final byte scalar) {
+    public Vector2F multiply(final byte scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D multiply(final short scalar) {
+    public Vector2F multiply(final short scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D multiply(final int scalar) {
+    public Vector2F multiply(final int scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D multiply(final long scalar) {
+    public Vector2F multiply(final long scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D multiply(final float scalar) {
+    public Vector2F multiply(final float scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D multiply(final double scalar) {
+    public Vector2F multiply(final double scalar) {
         x = (x * scalar);
         y = (y * scalar);
-        z = (z * scalar);
 
         return this;
     }
@@ -155,55 +141,49 @@ public class Vector3D implements VectorVector3Operations<Vector3D> {
     /*================================ Division ================================*/
 
     @Override
-    public Vector3D divide(final byte scalar) {
+    public Vector2F divide(final byte scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D divide(final short scalar) {
+    public Vector2F divide(final short scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D divide(final int scalar) {
+    public Vector2F divide(final int scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D divide(final long scalar) {
+    public Vector2F divide(final long scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D divide(final float scalar) {
+    public Vector2F divide(final float scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }
 
     @Override
-    public Vector3D divide(final double scalar) {
+    public Vector2F divide(final double scalar) {
         x = (x / scalar);
         y = (y / scalar);
-        z = (z / scalar);
 
         return this;
     }