+++ core/src/main/java/dev/redstudio/redcore/math/vectors/VectorVector2Operations.java
@@ -6,12 +6,21 @@ import dev.redstudio.redcore.annotations.Experimental;
 @SuppressWarnings("unused")
 interface VectorVector2Operations<T> extends VectorPrimitiveOperations<T> {
 
+    // region Assignment
+    /*================================ Assignment ================================*/
+
+    T copy(final Vector2F vector);
+    T copy(final Vector2D vector);
+
+    // endregion
+
     // region Math
     /*================================ Math ================================*/
 
     // region Addition
     /*================================ Addition ================================*/
 
+    T add(final Vector2F vector);
     T add(final Vector2D vector);
 
     // endregion
@@ -19,6 +28,7 @@ interface VectorVector2Operations<T> extends VectorPrimitiveOperations<T> {
     // region Subtraction
     /*================================ Subtraction ================================*/
 
+    T subtract(final Vector2F vector);
     T subtract(final Vector2D vector);
 
     // endregion