+++ core/src/main/java/dev/redstudio/redcore/math/vectors/VectorVector3Operations.java
@@ -6,6 +6,14 @@ import dev.redstudio.redcore.annotations.Experimental;
 @SuppressWarnings("unused")
 interface VectorVector3Operations<T> extends VectorPrimitiveOperations<T> {
 
+    // region Assignment
+    /*================================ Assignment ================================*/
+
+    T copy(final Vector3F vector);
+    T copy(final Vector3D vector);
+
+    // endregion
+
     // region Math
     /*================================ Math ================================*/
 
@@ -13,12 +21,14 @@ interface VectorVector3Operations<T> extends VectorPrimitiveOperations<T> {
     /*================================ Addition ================================*/
 
     T add(final Vector3D vector);
+    T add(final Vector3F vector);
 
     // endregion
 
     // region Subtraction
     /*================================ Subtraction ================================*/
 
+    T subtract(final Vector3F vector);
     T subtract(final Vector3D vector);
 
     // endregion